refactor: extract CarouselPanel and ToolsPanel from SimScreenView
SimScreenView was a ~1150-line orchestrator.  Two identifiable
sub-concerns with clear constructor boundaries are now separate
Node subclasses:

CarouselPanel (112 lines)
  • Owns the ComponentCarousel, PageControl, GridScaleIndicatorNode.
  • Registers the visibleBoundsProperty listener that pins the
    toolbox to the left safe-area edge.
  • Registers the gridSizeProperty listener that rebuilds the
    scale indicator when spacing changes.
  • Exposes `carousel` and `pageControl` for z-order and PDOM use.

ToolsPanel (348 lines)
  • Owns MeasuringTapeNode, ProtractorNode, all 11 toggle
    Checkboxes, the ray-density NumberControl, and the AccordionBox.
  • Manages the bidirectional mode sync (extended / images /
    observer ↔ modeProperty) internally.
  • Pins the accordion to the upper-right safe area via
    visibleBoundsProperty.
  • Exposes measuringTapeNode, protractorNode, accordionBox,
    and the three boolean properties needed by the keyboard-
    shortcut handler and SVG exporter.
  • Provides reset() so the ResetAllButton listener in
    SimScreenView stays a one-liner.

SimScreenView constructor shrank from 1149 → 830 lines.
Observer-node visibility is now driven directly from
model.scene.modeProperty.link() rather than through
the mode-sync closure, removing the coupling between
ToolsPanel and ObserverNode.
